incorporate functionality from #11
